moves page views into pages folder
