+++ src/routes.jsx
@@ -7,8 +7,8 @@ const {
 } = require('react-router');
 
 const App = require('./components/App');
-const Home = require('./components/Page/Home');
-const About = require('./components/Page/About');
+const Home = require('./pages/home');
+const About = require('./pages/about');
 
 module.exports = (
   <Route name="app" path="/" handler={App}>
